+++ HISTORY.md
@@ -3,6 +3,12 @@
   * Added `Multi Label Soft Margin Loss` loss function for neural networks
    (#2345).
 
+  * Updated terminal state for Pendulum environment (#2354).
+
+  * Added `EliSH` activation function (#2323).
+
+  * Add L1 Loss function (#2203).
+
 ### mlpack 3.3.0
 ###### 2020-04-07
   * Templated return type of `Forward function` of loss functions (#2339).